Performance enhancement
As we monitor more builds and keep longer history, the first landing
page becomes slow. The reason is that we are trying to prepare all data
regardless it will show up or not. In the landing page, due to table
pagination we only show the latest 5 builds from each pipeline. The rest
does not need to be queried until the user clicks on the next page.

This PR updates the logic to simplify TopLevelBuilds.jsx. We separate it
to contain React components. Each table/pipeline is one component. Each
component keep tracks its own pagination. Query is based on the table
pagination.

Signed-off-by: lanxia <lan_xia@ca.ibm.com>