Support registration and verification
Users can use api/auth/register to register their info. With the
registration, TRSS will generate a token using JWT (sets to expire in
24hrs). Users can send TRSS test results with this token using
api/auth/verify. TRSS will verify the token and return user info.

This is the initial version of the registration and verification. We
will work on the data insertion and display in the future PRs.

Signed-off-by: lanxia <lan_xia@ca.ibm.com>