Process large output
- Due to 1G string limit, we cannot store ouptut > 1G. Using http head
request to get the output size in getSize(). If the size > 1G, only
process the last 0.7G. This is a temporary fix, so that TRSS can proceed
and continue processing other builds.

- instead of using Jenkins nodejs API, we change to use `got` to query.
This will allow us to control the query precisely.

- increase regular query timeout from 2secs to 30secs and reduce the
retry from 5 times to 3 times for Jenkins nodejs APIs. It is better to
retry less often with longer timeout

- increase node max memory to 6G


Signed-off-by: lanxia <lan_xia@ca.ibm.com>