Remove build result check when set status to streaming (#378)
In the original code, we added the build result check to ensure we do not set streaming when the build is completed.
In this case, build can be processed directly. We use build result != null to determine if a build is completed or not.
However, we noticed a running build could have result != null. Because this build has build result, TRSS thinks the
the build is complete and tries to process the build in the static way (without streaming). At a later stage, when
TRSS notices this build is still running, it can do nothing but wait for the build to complete. From the user end, it
looks like the build is set as streaming but no data show up until the whole pipeline is completed.

Signed-off-by: lanxia <lan_xia@ca.ibm.com>