Update build retry logic
When querying a build, TRSS will retry 3 times.
In the case of an error, the build will move back to the queue if the return code from Jenkins API is not 404.
If the return code is 404, x-jenkins header is checked to ensure the error code is from Jenkins (not nginx).
In this case, it means the build has an invalid url or is expired. retry function returns 404 and TRSS will stop processing this build.
Otherwise, the error will be ignored (network issue) and the build will move back to the queue.
return code 404 only if the Jenkins server returns 404 for the build (i.e., invalid url, expired build, etc)

- sleep is removed to increase efficiency and reliability. We will check the header in the response to determine the invalid/expired build case.
- test API getJenkinsBuildInfo is added
- fix log issue

related: https://github.com/adoptium/aqa-test-tools/issues/852

Signed-off-by: Lan Xia <Lan_Xia@ca.ibm.com>